데이터 관련 Chatting 키워드를 Script 키워드로 변경 및 entity 형태 추가
